Fix error handling when there is no response
Even though the exception name suggests otherwise, sometimes the `response` in
the exception is nil. Trying to call `body` on it, then results in the following
exception:

```
NoMethodError: undefined method `body' for nil:NilClass

        error_msg = JSON.parse(e.response.body)['error']['message']
```

So, it seems that we need to check for the existence of a body, before we try to
parse it.